+++ reconstruction/coor_plot.h
@@ -22,9 +22,9 @@ class coor_plot {
 public:
     void create_figure();
     void show();
-    void plot_constraint_line(double k, double bias);
-    void plot_ellipse(vector<double> x);
-//    void plot_3d_model(vector<vector<double>> X_arr, vector<double> z_arr);
+    void plot_constraint_line(float k, float bias);
+    void plot_ellipse(vector<float> x);
+//    void plot_3d_model(vector<vector<float>> X_arr, vector<float> z_arr);
     void save(string filename);
     void close();
 
@@ -35,18 +35,18 @@ private:
 
 namespace coor_plot_3d {
     // 函数
-    void set_point_cloud(vector<vector<double>> X_3D, vector<vector<double>> Y_3D, vector<vector<double>> Z_3D);
+    void set_point_cloud(vector<vector<float>> X_3D, vector<vector<float>> Y_3D, vector<vector<float>> Z_3D);
     void render_screen();
     void setupRenderingState();
     void changeSize(GLint w, GLint h);
     void keyFunc(int key, int x, int y);
     void mouseFunc(int buttion, int state, int x, int y);
     void mouseMotion(int x, int y);
-    void plot_3d_model(vector<vector<double>> X, vector<vector<double>> Y, vector<vector<double>> Z);
+    void plot_3d_model(vector<vector<float>> X, vector<vector<float>> Y, vector<vector<float>> Z);
 
     void render_screen_finger_vein();
-    void set_finger_vein_point_cloud(vector<vector<double>> X, vector<vector<double>> Y, vector<vector<double>> Z, vector<vector<unsigned char>> texture);
-    void plot_3d_finger_vein_model(vector<vector<double>> X, vector<vector<double>> Y, vector<vector<double>> Z, vector<vector<unsigned char>> texture);
+    void set_finger_vein_point_cloud(vector<vector<float>> X, vector<vector<float>> Y, vector<vector<float>> Z, vector<vector<unsigned char>> texture);
+    void plot_3d_finger_vein_model(vector<vector<float>> X, vector<vector<float>> Y, vector<vector<float>> Z, vector<vector<unsigned char>> texture);
 };
